making our way downtown..
