+++ crates/sui-graphql-rpc/src/types/checkpoint.rs
@@ -4,13 +4,11 @@
 use crate::context_data::sui_indexer_reader::PgManager;
 
 use super::{
-    base64::Base64, digest::Digest, end_of_epoch_data::EndOfEpochData, epoch::Epoch,
+    base64::Base64, end_of_epoch_data::EndOfEpochData, epoch::Epoch,
     gas::GasCostSummary,
 };
 use async_graphql::*;
-use sui_indexer::models_v2::checkpoints::StoredCheckpoint;
 
-use crate::error::Error;
 
 #[derive(InputObject)]
 pub(crate) struct CheckpointId {
@@ -31,35 +29,10 @@ pub(crate) struct Checkpoint {
     pub network_total_transactions: Option<u64>,
     pub rolling_gas_summary: Option<GasCostSummary>,
     #[graphql(skip)]
-    pub epoch_id: u64,
+    pub epoch: Epoch,
     pub end_of_epoch: Option<EndOfEpochData>,
     // transactionConnection(first: Int, after: String, last: Int, before: String): TransactionBlockConnection
     // address_metrics: AddressMetrics,
-}
-
-impl TryFrom<StoredCheckpoint> for Checkpoint {
-    type Error = Error;
-    fn try_from(c: StoredCheckpoint) -> Result<Self, Self::Error> {
-        Ok(Self {
-            digest: Digest::try_from(c.checkpoint_digest)?.to_string(),
-            sequence_number: c.sequence_number as u64,
-            validator_signature: Some(c.validator_signature.into()),
-            previous_checkpoint_digest: c
-                .previous_checkpoint_digest
-                .map(|d| Digest::try_from(d).map(|digest| digest.to_string()))
-                .transpose()?,
-            live_object_set_digest: None,
-            network_total_transactions: Some(c.network_total_transactions as u64),
-            rolling_gas_summary: Some(GasCostSummary {
-                computation_cost: c.computation_cost as u64,
-                storage_cost: c.storage_cost as u64,
-                storage_rebate: c.storage_rebate as u64,
-                non_refundable_storage_fee: c.non_refundable_storage_fee as u64,
-            }),
-            epoch_id: c.epoch as u64,
-            end_of_epoch: None,
-        })
-    }
 }
 
 #[ComplexObject]
@@ -67,7 +40,7 @@ impl Checkpoint {
     async fn epoch(&self, ctx: &Context<'_>) -> Result<Option<Epoch>> {
         let result = ctx
             .data_unchecked::<PgManager>()
-            .fetch_epoch_strict(self.epoch_id)
+            .fetch_epoch_strict(self.epoch)
             .await?;
         Ok(Some(Epoch::from(result)))
     }
