+++ crates/sui-graphql-rpc/src/types/transaction_block.rs
@@ -54,53 +54,6 @@ impl From<SuiTransactionBlockResponse> for TransactionBlock {
     }
 }
 
-impl TryFrom<StoredTransaction> for TransactionBlock {
-    type Error = Error;
-
-    fn try_from(tx: StoredTransaction) -> Result<Self, Self::Error> {
-        // TODO (wlmyng): Split the below into resolver methods
-        let digest = Digest::try_from(tx.transaction_digest.as_slice())?;
-
-        let sender_signed_data: SenderSignedData =
-            bcs::from_bytes(&tx.raw_transaction).map_err(|e| {
-                Error::Internal(format!(
-                    "Can't convert raw_transaction into SenderSignedData. Error: {e}",
-                ))
-            })?;
-
-        let sender = Address {
-            address: SuiAddress::from_array(
-                sender_signed_data
-                    .intent_message()
-                    .value
-                    .sender()
-                    .to_inner(),
-            ),
-        };
-
-        let gas_input = GasInput::from(sender_signed_data.intent_message().value.gas_data());
-        let effects: TransactionEffects = bcs::from_bytes(&tx.raw_effects).map_err(|e| {
-            Error::Internal(format!(
-                "Can't convert raw_effects into TransactionEffects. Error: {e}",
-            ))
-        })?;
-        let effects = match SuiTransactionBlockEffects::try_from(effects) {
-            Ok(effects) => Ok(Some(TransactionBlockEffects::from(&effects))),
-            Err(e) => Err(Error::Internal(format!(
-                "Can't convert TransactionEffects into SuiTransactionBlockEffects. Error: {e}",
-            ))),
-        }?;
-
-        Ok(Self {
-            digest,
-            effects,
-            sender: Some(sender),
-            bcs: Some(Base64::from(&tx.raw_transaction)),
-            gas_input: Some(gas_input),
-        })
-    }
-}
-
 #[ComplexObject]
 impl TransactionBlock {
     async fn digest(&self) -> String {
